requirements.txt
openai
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
langchain-chroma
pydantic
pandas
tiktoken
pymupdf
beautifulsoup4


